pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "genehomology"
version = "0.1.0"
description = "跨物种同源基因序列比对工具"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "GeneHomology Team", email = "your-email@example.com"}
]
keywords = ["bioinformatics", "gene", "homology", "blast", "sequence-analysis"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
requires-python = ">=3.7"
dependencies = [
    "requests>=2.25.0",
    "biopython>=1.78",
]

[project.urls]
Homepage = "https://github.com/yourusername/genehomology"
Repository = "https://github.com/yourusername/genehomology"

[project.scripts]
genehomology = "genehomology.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["genehomology*"]
exclude = ["tests*"] 